# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/discussions/lesson/{lessonID}": {
            "get": {
                "description": "ListController for Discussions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discussions"
                ],
                "summary": "ListController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "lessonID",
                        "name": "lessonID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.PaginatedResponse-discussions_DiscussionResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateController for Discussions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discussions"
                ],
                "summary": "CreateController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "lessonID",
                        "name": "lessonID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/discussions.CreateDiscussionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-discussions_Discussion"
                        }
                    }
                }
            }
        },
        "/api/v1/discussions/replies/{id}": {
            "get": {
                "description": "ListRepliesController for Discussions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discussions"
                ],
                "summary": "ListRepliesController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.PaginatedResponse-discussions_DiscussionResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/discussions/{id}": {
            "delete": {
                "description": "DeleteController for Discussions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discussions"
                ],
                "summary": "DeleteController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-utils_DeleteResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "UpdateController for Discussions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discussions"
                ],
                "summary": "UpdateController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/discussions.UpdateDiscussionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-discussions_Discussion"
                        }
                    }
                }
            }
        },
...
        "/api/v1/enrollments/manual/course/{courseID}": {
            "post": {
                "description": "CreateController for Enrollments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enrollments"
                ],
                "summary": "CreateController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "courseID",
                        "name": "courseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enrollments.ManualEnrollRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-enrollments_Enrollment"
                        }
                    }
                }
            }
        },
...
        "/api/v1/feedbacks/course/{courseID}": {
            "post": {
                "description": "CreateController for Feedbacks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feedbacks"
                ],
                "summary": "CreateController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "courseID",
                        "name": "courseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/feedbacks.CreateFeedbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-feedbacks_Feedback"
                        }
                    }
                }
            }
        },
        "/api/v1/feedbacks/{id}": {
            "delete": {
                "description": "DeleteController for Feedbacks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feedbacks"
                ],
                "summary": "DeleteController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-utils_DeleteResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/feedbacks/{id}/pin": {
            "patch": {
                "description": "UpdateController for Feedbacks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feedbacks"
                ],
                "summary": "UpdateController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-feedbacks_Feedback"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "discussions.CreateDiscussionRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 5000,
                    "minLength": 1
                },
                "parent_id": {
                    "type": "string"
                }
            }
        },
        "discussions.Discussion": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "course_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "depth": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "reply_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.UserInfo"
                }
            }
        },
        "discussions.DiscussionResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "depth": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "reply_count": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/models.UserInfo"
                }
            }
        },
        "discussions.UpdateDiscussionRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 5000,
                    "minLength": 1
                }
            }
        },
...
        "enrollments.Enrollment": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "completion_percent": {
                    "type": "number"
                },
                "course_id": {
                    "type": "string"
                },
                "enrolled_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_accessed_lesson_id": {
                    "type": "string"
                },
                "revoked": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "enrollments.ManualEnrollRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "feedbacks.CreateFeedbackRequest": {
            "type": "object",
            "required": [
                "rating"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1
                }
            }
        },
        "feedbacks.Feedback": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "course": {
                    "$ref": "#/definitions/models.CourseInfo"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_pinned": {
                    "type": "boolean"
                },
                "rating": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/models.UserInfo"
                }
            }
        },
...
        "models.CourseInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.UserInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "utils.DeleteResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "utils.PaginatedData-discussions_DiscussionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/discussions.DiscussionResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "utils.PaginatedResponse-discussions_DiscussionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/utils.PaginatedData-discussions_DiscussionResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.SwaggerResponse-discussions_Discussion": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/discussions.Discussion"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.SwaggerResponse-enrollments_Enrollment": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/enrollments.Enrollment"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.SwaggerResponse-feedbacks_Feedback": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/feedbacks.Feedback"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.SwaggerResponse-utils_DeleteResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/utils.DeleteResponse"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}